+++ src/app/components/elements/Socail.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { SnsStype } from 'app/components/modules/footer/contact/Contact';
+import { SnsStype } from 'app/components/modules/Footer';
 
 const Socail:React.FC<SnsStype> = (sns):React.ReactElement =>  {
   return (
