--- src/app/components/modules/footer/copyRight/CopyRight.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-
-const CopyRight:React.FC = (): React.ReactElement => {
-  return (
-    <div className="CopyRight">
-      <p className="content">TheBox Company © 2022. All Rights Reserved</p>
-    </div>
-  );
-};
-export default CopyRight;
