--- src/app/components/modules/reputation/ReputationType.ts
@@ -1,6 +1,0 @@
-
-export type ReputationType = {
-    img : string ;
-    title : string;
-    decs : string;
-}
