--- src/app/components/modules/services/ServicesType.ts
@@ -1,5 +1,0 @@
-export type ServicesType = {
-    id : number ;
-    icon : string ;
-    name : string ;
-}
